`build` and `buildx` commands removed
There's no added value as the outcome of the rendering process
doesn't depend on the subsequent build process.
So we don't need to run the build / buildx commands in here.
This reduces complexity and allows to properly specify all
desired parameters for the subsequent `docker build` and
`docker buildx build` calls freely.
So the replacement for what was so far:

    docker-template build 'your-tag'

would now be:

    docker-template render 'your-tag' && \
    docker build -t 'your-tag' .